test: fix potential overflow

+++ test/tiffRead.cpp
@@ -17,7 +17,7 @@ bool testLibTiffImage(const std::string& filename) {
     std::cout << "-------------------------------------------------" << std::endl;
     std::cout << "Test libtiff images: " << path << std::endl;
     if (const auto dermTiff = ldt::io::OpenTIFF(path); dermTiff.isValid) {
-        Image raster(dermTiff.width * dermTiff.height);
+        Image raster(static_cast<size_t>(dermTiff.width) * dermTiff.height);
         ldt::Pencil pencil;
         return ldt::io::ReadPage(path, 0, reinterpret_cast<uint32_t*>(raster.data()), &pencil);
     }
@@ -29,7 +29,7 @@ bool testDermTiffImage(const std::string& filename, const std::vector<ldt::Penci
     std::cout << "-------------------------------------------------" << std::endl;
     std::cout << "Test DermAnnotation TIFF images: " << path << std::endl;
     if (const auto dermTiff = ldt::io::OpenTIFF(path); dermTiff.isValid) {
-        Image raster(dermTiff.width * dermTiff.height);
+        Image raster(static_cast<size_t>(dermTiff.width) * dermTiff.height);
         // Read original image
         {
             if (!ldt::io::ReadOriginalImage(path, reinterpret_cast<uint32_t*>(raster.data()))) {
